Add recycler adapter

+++ app/src/main/java/com/example/burgershub/data/model/BurgerDTO.kt
@@ -6,6 +6,6 @@ data class BurgerDTO(
   val images: List<ImageDTO>?,
   val ingredients: List<IngredientDTO>?,
   val name: String?,
-  val price: Double?,
+  val price: Float?,
   val veg: Boolean?,
 )
